Renamed AST lib to avoid name collisions
